Rep and more refactoring + new annotation
