+++ src/main/java/org/example/service/UserRegistrationServiceEncrypted.java
@@ -1,14 +1,18 @@
 package org.example.service;
 
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 import org.example.users.User;
 import org.example.repository.UserRepository;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
+@ApplicationScoped
 public class UserRegistrationServiceEncrypted implements UserRegistrationService {
   private final UserRepository userRepository;
 
+  @Inject
   public UserRegistrationServiceEncrypted(UserRepository userRepository) {
     this.userRepository = userRepository;
   }
